Fix clippy warning: remove unnecessary cast
SIGHASH_ALL and SIGHASH_FORKID are already u32, so casting their
bitwise OR result to u32 is unnecessary.

+++ src/bsv/tx_signature.rs
@@ -245,7 +245,7 @@ mod tests {
         // Scope should match (only low byte)
         assert_eq!(
             decoded.scope() & 0xff,
-            (SIGHASH_ALL | SIGHASH_FORKID) as u32
+            SIGHASH_ALL | SIGHASH_FORKID
         );
     }
